change firebase authentication to class
