+++ api/app/routers/misptags.py
@@ -3,7 +3,7 @@ from fastapi import Query as QueryParameter
 from sqlalchemy.orm import Session
 
 from app import models, persistence, schemas
-from app.auth import get_current_user
+from app.account import get_current_user
 from app.database import get_db
 
 router = APIRouter(prefix="/misp_tags", tags=["misp_tags"])
